Move all objects to same module
